% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_split.R
\name{split_panel}
\alias{split_panel}
\title{Split panel data}
\usage{
split_panel(dat, unit_col, treatment_column, num_post = NULL)
}
\arguments{
\item{dat}{A data frame with columns as time periods and rows as units. Can contain a column for unit names but otherwise should only contain values.}

\item{unit_col}{A character vector of unit names. Should match the unit column if it exists.}

\item{treatment_column}{A character object of the first treated period's variable name.}

\item{num_post}{The amount of post-treatment periods you want to keep. Defaults to NULL which keeps all.}
}
\value{
List of each direction's pre- and post- treatment matrices.
}
\description{
Splits panel data (columns as time and rows as units) into horizontal and vertical pre- and post- treatment matrices.
}
